Use name as title (#37)
Claude code appears to have a bug where it uses the prompt's `title`
instead of its `name`, and chokes on it if it contains whitespace
characters. This changes uses the existing `name` for the `title` as
well, making them consistently use snake_case, and work around the
problem for now.

More detail in
[slack](https://iobeam.slack.com/archives/C099S47CSRX/p1761078168570189)